requirements.txt
json
neo4j==5.14.1
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.17
pydantic==2.5.2
python-dotenv==1.0.1
tiktoken==0.5.2
typing-inspect==0.9.0
streamlit
google-generativeai==0.4.1
openai==1.12.0
sentence-transformers